отдельный класс UserDict
